# Added Rcpp function for likelihood_allele()