--- a/DESCRIPTION
+++ b/DESCRIPTION
@@ -44,7 +44,7 @@ Imports:
 Suggests: 
     testthat (>= 3.0.0)
 Config/testthat/edition: 3
-LinkingTo: Rcpp
+LinkingTo: Rcpp, RcppArmadillo
 NeedsCompilation: yes
 SystemRequirements: C++11
 Author: Teng Gao [cre, aut], Hirak Sarkar [aut], Evan Biederstedt [aut], Peter Kharchenko [aut]