--- a/src/likelihood_allele.cpp
+++ b/src/likelihood_allele.cpp
@@ -1,5 +1,4 @@
 
-#include <Rcpp.h>
 #include <RcppArmadillo.h>
 
 using namespace Rcpp;