* For the slow part of the function, I added `likelihood_allele_compute()` (which could use a better name....log likelihood calculations)


Note: Re-writing the entire function actually was a bit clumsy; I didn't test performance, but I suspect given some of the data structures I ended up using, it's just quicker to use R in these parts of the code (which just called C anyways). Simply converting everything to Rcpp/C++ isn't always quicker I've found (which is backed up in the docs + StackOverflow conversations)

Please check it works. I was careful to use the example as inputs when writing the function, but it's always possible there are mistakes.