@teng-gao 2022-02-05
Thanks Evan! This looks like the correct approach. Will test myself after Tuesday. At the mean time could you get some performance metrics compared to the old R function? I'd test it on the outer loop too (`approx_theta_post`)