@evanbiederstedt 2022-02-06
RE: performance

I used [tictoc](https://cran.r-project.org/web/packages/tictoc/index.html), and just ran the example; I didn't isolate the individual function. 

```
library(tictoc)
library(numbat)

bulk = get_bulk(
    count_mat = count_mat_ATC2,
    df_allele = df_allele_ATC2,
    lambdas_ref = ref_hca,
    gtf_transcript = gtf_hg38,
    genetic_map = genetic_map_hg38
)


tic()
for (i in 1:20) {
  bulk = analyze_bulk(bulk, t = 1e-5)
}
toc()
```

Which runs as we'd expect, i.e. 

```
INFO [2022-02-05 22:33:38] diploid regions: 5a,7a,7c,19b
Retesting CNVs..
Finishing..
... etc
```

Before Rcpp:
* Total time: 1576.241 sec elapsed. That's an average of 1576.241 / 20 = 78.81 seconds. 

After Rcpp (i.e. this PR): 
* Total time: 359.129 sec elapsed. That's an average of 359.129 / 20 = 17. 96 seconds

1576.241 / 359.129 = 4.39

Let's call it a 4-5x speedup.