@teng-gao 2022-02-07
That's awesome - did you compare if the results are exactly the same? I would do it both for the output of `analyze_bulk` and the entire `run_numbat` function.