@teng-gao 2022-02-07
Hmm, I tested this on the same example. There was indeed a speedup but the results are very different. Please investigate. Figures produced by `bulk %>% plot_bulks()`
![image](https://user-images.githubusercontent.com/13375875/152847808-6e03ecef-f495-4802-85a1-32471be4073b.png)
![image](https://user-images.githubusercontent.com/13375875/152847895-b5451d60-2c35-4cb3-9f31-41e63a4df941.png)