@evanbiederstedt 2022-02-08
@teng-gao This looks better---there was a mistake here: https://github.com/kharchenkolab/Numbat/pull/11/commits/85232f1419832d6cf3aeedd2e8be360e385fda29

After fix:
![fix](https://user-images.githubusercontent.com/6968193/152906221-055dbab4-2b0d-4861-9bb9-7ef9adee60c5.jpg)


The numerical results are not 100% the same though I've noticed; there's a slight difference. 
* Original likelihood calculated by `likelihood_allele()`:  -736.1658
* After Rcpp: -739.1069

There's some numeric discrepancy due to the C++ types vs R types. I thought the `logSumExp()` written in C++ gave the same values as the `matrixStats::logSumExp` but that could be the source of the slight difference as well. It's possible we'll never get it 100% numerically the same...