@teng-gao 2022-02-08
Looks much better indeed. I'm a bit surprised that the total likelihood can be that different as these should be simple numerical additions. I'd even suspect an off by one error causing this e.g. adding one more or less datapoint to the total likelihood