@evanbiederstedt 2022-02-08
> I'd even suspect an off by one error causing this e.g. adding one more or less datapoint to the total likelihood

This seems very likely, as I could have screwed up the logic between R 1-indexing and C++ 0-indexing....let me have another look...

I've tried the `logSumExp()` function----exactly the same as `matrixStats::logSumExp()`:

```
> foo
      [,1]  [,2]
[1,] 1e+00 1e-10
[2,] 1e-10 1e+00
l> logSumExp(foo)
[1] 2.006409
> matrixStats::logSumExp(foo)
[1] 2.006409
> foobar = hmm$Pi[[5]]
> foobar
            [,1]        [,2]
[1,] 0.997986797 0.002013203
[2,] 0.002013203 0.997986797
> logSumExp(foobar)
[1] 2.00548
> matrixStats::logSumExp(foobar)
[1] 2.00548
```