@evanbiederstedt 2022-02-08
> I'd even suspect an off by one error causing this e.g. adding one more or less datapoint to the total likelihood

Yes, I think I found it. 

This:
```
 if (i > 1) {
      logphi <- sapply(1:m, function(j) matrixStats::logSumExp(logphi + logPi[[i]][,j]))
}
```

should be `i > 0` in C++

Now we have the following:

* Original likelihood calculated by likelihood_allele(): -736.1658
* After Rcpp: -736.8074

Much better