@teng-gao 2022-02-09
This would be a great test in `testthat/` - would you have a chance to add?