@evanbiederstedt 2022-02-09
> This would be a great test in testthat/ - would you have a chance to add?

Done.

Added tests, they pass. 

Let's wrap this PR up. We can try to speed up in the future. 

BTW, I exported the function `logSumExp()`. We could replace `matrixStats::logSumExp()` with this (and even remove the package `matrixStats` from the dependencies).